updated perf test

+++ autothon/features/environment.py
@@ -46,6 +46,7 @@ def deleteFileWithContent(content):
     entries = os.listdir('output/allure-reports/')
     for entry in entries:
         # print(entry)
-        file1 = open('output/allure-reports/' + entry, 'r')
-        if content in file1.read():
-            os.remove('output/allure-reports/' + entry)
+        if "." in entry:
+            file1 = open('output/allure-reports/' + entry, 'r')
+            if content in file1.read():
+                os.remove('output/allure-reports/' + entry)
